PP-7893 Use new designs for API key pages
Create a separate template `revoked-keys.njk` for the revoked keys page
as it has significantly diverged now. Continue using `_keys.njk` sub
template but move content specific for active keys into `index.njk`.

Modify tests to be integration tests rather than functional tests so
they're faster and easier to debug!
