+++ app/controllers/api-keys/get-create.controller.js
@@ -2,7 +2,7 @@
 
 const { response } = require('../../utils/response.js')
 
-module.exports = (req, res) => {
+module.exports = function showCreateApiKey (req, res) {
   const accountId = req.account.gateway_account_id
   response(req, res, 'api-keys/create', { 'account_id': accountId })
 }
