+++ app/views/api-keys/_keys.njk
@@ -1,4 +1,4 @@
-<h2 id="{{header}}" class="govuk-heading-m govuk-!-margin-top-6">
+<h2 class="govuk-heading-m govuk-!-margin-top-6">
   {% if tokens.length %}
     {% if tokens_singular %}
       There is 1 {{token_state}} API key
@@ -10,35 +10,6 @@
   {% endif %}
 </h2>
 
-{% if active %}
-  {% set apiKeyLinkRoute = routes.account.apiKeys.revoked %}
-  {% set apiKeyLinkText = 'revoked' %}
-  {% set apiKeyLinkId = 'revoked-keys-link' %}
-{% else %}
-  {% set apiKeyLinkRoute = routes.account.apiKeys.index %}
-  {% set apiKeyLinkText = 'active' %}
-  {% set apiKeyLinkId = 'active-keys-link' %}
-{% endif %}
-
-<div class="govuk-body">
-  <a class="govuk-link" href="{{ formatAccountPathsFor(apiKeyLinkRoute, currentGatewayAccount.external_id) }}" id="{{apiKeyLinkId}}">View {{apiKeyLinkText}} keys</a>
-</div>
-
-{% if permissions.tokens_create %}
-  {% if active %}
-    {{
-      govukButton({
-        text: "Create a new API key",
-        href: formatAccountPathsFor(routes.account.apiKeys.create, currentGatewayAccount.external_id),
-        classes: "generate-key",
-        attributes: {
-          id: "create-api-key"
-        }
-      })
-    }}
-  {% endif %}
-{% endif %}
-
 {% if tokens.length %}
 <div class="key-list">
   {% for token in tokens %}
@@ -46,3 +17,4 @@
   {% endfor %}
 </div>
 {% endif %}
+
